+++ src/app/professor/[name]/_component/personal/ProfessorPersonalView.tsx
