+++ src/app/professor/[name]/_component/personal/ProfessorProfileCard.tsx
