+++ src/app/professor/[name]/_component/seminar/SeminarAppealPoint.tsx
