+++ src/app/professor/[name]/_component/seminar/SeminarSchedule.tsx
